Optimize: remove redundant judgment
